digitalSTROM: - Added powerconsumption to IDeviceInterface, thus it's possible   to ask a group, a zone, a device and a set for it's combined   power consumption in mW. - Began implementing the common JSON API

+++ unix/ds485.cpp
@@ -477,7 +477,7 @@ namespace dss {
               }
               keep = true;
             } else {
-              cout << "&&&&&&&&&& unknown frame id: " << cmdFrame->GetCommand() << endl;
+              cout << "&&&&&&&&&& unknown frame id: " << (int)cmdFrame->GetCommand() << endl;
             }
             if(keep) {
               // put in into the received queue
